+++ src/Domains/comments/CommentRepository.js
@@ -18,14 +18,6 @@ class CommentRepository {
   async getCommentsByThreadId(threadId) {
     throw new Error('COMMENT_REPOSITORY.METHOD_NOT_IMPLEMENTED');
   }
-
-  async addLikeById({ id, userId }) {
-    throw new Error('COMMENT_REPOSITORY.METHOD_NOT_IMPLEMENTED');
-  }
-
-  async deleteLikeById({ id, userId }) {
-    throw new Error('COMMENT_REPOSITORY.METHOD_NOT_IMPLEMENTED');
-  }
 }
 
 module.exports = CommentRepository;
